Split out ArtifactRepresentation.
